# Improve performance of the FastlyLogProcessor job