--- a/app/models/gem_download.rb
+++ b/app/models/gem_download.rb
@@ -55,19 +55,26 @@ class GemDownload < ApplicationRecord
       end
 
       return if updates_by_version.empty?
-      updates_by_version.each_value do |version, version_count|
-        updates_by_gem[version.rubygem_id] ||= 0
-        updates_by_gem[version.rubygem_id] += version_count
-      end
 
-      updates_by_version.values.sort_by { |v, _| v.id }.each do |version, count|
-        # Gem version count
-        increment(count, rubygem_id: version.rubygem_id, version_id: version.id)
+      total_count = 0
+      updates_by_version.each_value.each_slice(1_000) do |versions|
+        rubygem_ids = []
+        version_ids = []
+        downloads = []
+        versions.each do |(version, version_count)|
+          updates_by_gem[version.rubygem_id] ||= 0
+          updates_by_gem[version.rubygem_id] += version_count
+
+          total_count += version_count
+
+          rubygem_ids << version.rubygem_id
+          version_ids << version.id
+          downloads << version_count
+        end
+        increment_versions(rubygem_ids, version_ids, downloads)
       end
 
       update_gem_downloads(updates_by_gem)
-
-      total_count = updates_by_gem.values.sum
 
       # Total count
       increment(total_count, rubygem_id: 0, version_id: 0)
@@ -97,12 +104,24 @@ class GemDownload < ApplicationRecord
       Rails.logger.debug { "ES update: #{updates_by_gem} has failed: #{e.message}" }
     end
 
+    def increment_versions(rubygem_ids, version_ids, downloads)
+      query = <<~SQL.squish
+        count = #{quoted_table_name}.count + updates_by_gem.downloads
+        FROM
+          (SELECT UNNEST(ARRAY[?]) AS r_id, UNNEST(ARRAY[?]) AS v_id, UNNEST(ARRAY[?]) AS downloads) AS updates_by_gem
+        WHERE #{quoted_table_name}.rubygem_id = updates_by_gem.r_id AND #{quoted_table_name}.version_id = updates_by_gem.v_id
+      SQL
+      update_all([query, rubygem_ids, version_ids, downloads])
+    end
+
     def increment_rubygems(rubygem_ids, downloads)
-      query = "UPDATE gem_downloads SET count = gem_downloads.count + updates_by_gem.downloads
+      query = <<~SQL.squish
+        count = #{quoted_table_name}.count + updates_by_gem.downloads
         FROM
           (SELECT UNNEST(ARRAY[?]) AS r_id, UNNEST(ARRAY[?]) AS downloads) AS updates_by_gem
-        WHERE gem_downloads.rubygem_id = updates_by_gem.r_id AND gem_downloads.version_id = 0;"
-      find_by_sql([query, rubygem_ids, downloads])
+        WHERE #{quoted_table_name}.rubygem_id = updates_by_gem.r_id AND #{quoted_table_name}.version_id = 0
+      SQL
+      update_all([query, rubygem_ids, downloads])
     end
 
     def downloads_by_gem(rubygem_ids)