Bulk increment version downloads
    Goes from N update queries to N/1000 for gem versions, which should significantly speed up the FastlyLogProcessor

Explicitly lock on the all gems GemDownload in FastlyLogProcessor
    Since every bulk_update will touch that row, pessimistic locking should help limit the amount of transaction conflicts concurrent processing causes

Example diff of `bulk_update` with seed data shows us going from 19 queries to 12, and the difference will be much starker with more than 8 versions...

Before:

```
irb(main):003:0> GemDownload.bulk_update(Version.all.select(:full_name, '0').pluck(:full_name, '0'))
  Version Pluck (0.4ms)  SELECT "versions"."full_name", 0 FROM "versions"
  Version Load (0.5ms)  SELECT "versions"."full_name", "versions"."rubygem_id", "versions"."id" FROM "versions" WHERE "versions"."full_name" IN ($1, $2, $3, $4, $5, $6, $7, $8)  [["full_name", "rubygem1-1.1.0.pre.2"], ["full_name", "rubygem1-1.0.0.pre.1"], ["full_name", "bundler-1.10.6"], ["full_name", "json-1.8.2"], ["full_name", "json-1.8.3-java"], ["full_name", "json-1.8.3"], ["full_name", "rubygem0-1.0.0"], ["full_name", "rubygem0-1.0.0-x86_64-darwin"]]
  GemDownload Load (1.1ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 8 AND "gem_downloads"."version_id" = 16) RETURNING *
  GemDownload Load (0.8ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 8 AND "gem_downloads"."version_id" = 17) RETURNING *
  GemDownload Load (0.9ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 9 AND "gem_downloads"."version_id" = 18) RETURNING *
  GemDownload Load (0.6ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 9 AND "gem_downloads"."version_id" = 19) RETURNING *
  GemDownload Load (0.6ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 10 AND "gem_downloads"."version_id" = 20) RETURNING *
  GemDownload Load (0.6ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 11 AND "gem_downloads"."version_id" = 21) RETURNING *
  GemDownload Load (1.9ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 11 AND "gem_downloads"."version_id" = 22) RETURNING *
  GemDownload Load (1.3ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 11 AND "gem_downloads"."version_id" = 23) RETURNING *
  GemDownload Load (1.6ms)  SELECT "gem_downloads".* FROM "gem_downloads" INNER JOIN "versions" ON "versions"."id" = "gem_downloads"."version_id" WHERE "versions"."latest" = $1 AND "versions"."platform" = $2 AND "gem_downloads"."rubygem_id" IN ($3, $4, $5, $6)  [["latest", true], ["platform", "ruby"], ["rubygem_id", 9], ["rubygem_id", 10], ["rubygem_id", 11], ["rubygem_id", 8]]
  Rubygem Load (0.3ms)  SELECT "rubygems".* FROM "rubygems" WHERE "rubygems"."id" = $1 LIMIT $2  [["id", 9], ["LIMIT", 1]]
  Version Load (0.7ms)  SELECT "versions".* FROM "versions" WHERE "versions"."rubygem_id" = $1 AND "versions"."latest" = $2 AND "versions"."platform" = $3 LIMIT $4  [["rubygem_id", 9], ["latest", true], ["platform", "ruby"], ["LIMIT", 1]]
  Version Load (0.6ms)  SELECT "versions".* FROM "versions" WHERE "versions"."rubygem_id" = $1 AND "versions"."latest" = $2 ORDER BY "versions"."number" DESC LIMIT $3  [["rubygem_id", 9], ["latest", true], ["LIMIT", 1]]
  Version Load (0.5ms)  SELECT "versions".* FROM "versions" WHERE "versions"."rubygem_id" = $1 ORDER BY "versions"."id" DESC LIMIT $2  [["rubygem_id", 9], ["LIMIT", 1]]
  GemDownload Load (0.3ms)  SELECT "gem_downloads".* FROM "gem_downloads" WHERE "gem_downloads"."version_id" = $1 LIMIT $2  [["version_id", 19], ["LIMIT", 1]]
  GemDownload Pluck (0.3ms)  SELECT "gem_downloads"."rubygem_id", "gem_downloads"."count" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" IN ($1, $2, $3, $4) AND "gem_downloads"."version_id" = $5 ORDER BY "gem_downloads"."rubygem_id" ASC  [["rubygem_id", 9], ["rubygem_id", 10], ["rubygem_id", 11], ["rubygem_id", 8], ["version_id", 0]]
  GemDownload Load (0.6ms)  UPDATE gem_downloads SET count = gem_downloads.count + updates_by_gem.downloads FROM (SELECT UNNEST(ARRAY[9,10,11,8]) AS r_id, UNNEST(ARRAY[0,0,0,0]) AS downloads) AS updates_by_gem WHERE gem_downloads.rubygem_id = updates_by_gem.r_id AND gem_downloads.version_id = 0;
  GemDownload Load (0.6ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 0 AND "gem_downloads"."version_id" = 0) RETURNING *
=> nil
```

After:

```
  Version Pluck (1.5ms)  SELECT "versions"."full_name", 0 FROM "versions"
  Version Load (0.8ms)  SELECT "versions"."full_name", "versions"."rubygem_id", "versions"."id" FROM "versions" WHERE "versions"."full_name" IN ($1, $2, $3, $4, $5, $6, $7, $8)  [["full_name", "rubygem1-1.1.0.pre.2"], ["full_name", "rubygem1-1.0.0.pre.1"], ["full_name", "bundler-1.10.6"], ["full_name", "json-1.8.2"], ["full_name", "json-1.8.3-java"], ["full_name", "json-1.8.3"], ["full_name", "rubygem0-1.0.0"], ["full_name", "rubygem0-1.0.0-x86_64-darwin"]]
  GemDownload Update All (2.0ms)  UPDATE "gem_downloads" SET count = "gem_downloads".count + updates_by_gem.downloads FROM (SELECT UNNEST(ARRAY[9,9,10,11,11,11,8,8]) AS r_id, UNNEST(ARRAY[19,18,20,23,21,22,16,17]) AS v_id, UNNEST(ARRAY[0,0,0,0,0,0,0,0]) AS downloads) AS updates_by_gem WHERE "gem_downloads".rubygem_id = updates_by_gem.r_id AND "gem_downloads".version_id = updates_by_gem.v_id
  GemDownload Load (1.0ms)  SELECT "gem_downloads".* FROM "gem_downloads" INNER JOIN "versions" ON "versions"."id" = "gem_downloads"."version_id" WHERE "versions"."latest" = $1 AND "versions"."platform" = $2 AND "gem_downloads"."rubygem_id" IN ($3, $4, $5, $6)  [["latest", true], ["platform", "ruby"], ["rubygem_id", 9], ["rubygem_id", 10], ["rubygem_id", 11], ["rubygem_id", 8]]
  Rubygem Load (1.2ms)  SELECT "rubygems".* FROM "rubygems" WHERE "rubygems"."id" = $1 LIMIT $2  [["id", 9], ["LIMIT", 1]]
  Version Load (0.5ms)  SELECT "versions".* FROM "versions" WHERE "versions"."rubygem_id" = $1 AND "versions"."latest" = $2 AND "versions"."platform" = $3 LIMIT $4  [["rubygem_id", 9], ["latest", true], ["platform", "ruby"], ["LIMIT", 1]]
  Version Load (0.5ms)  SELECT "versions".* FROM "versions" WHERE "versions"."rubygem_id" = $1 AND "versions"."latest" = $2 ORDER BY "versions"."number" DESC LIMIT $3  [["rubygem_id", 9], ["latest", true], ["LIMIT", 1]]
  Version Load (1.4ms)  SELECT "versions".* FROM "versions" WHERE "versions"."rubygem_id" = $1 ORDER BY "versions"."id" DESC LIMIT $2  [["rubygem_id", 9], ["LIMIT", 1]]
  GemDownload Load (0.4ms)  SELECT "gem_downloads".* FROM "gem_downloads" WHERE "gem_downloads"."version_id" = $1 LIMIT $2  [["version_id", 19], ["LIMIT", 1]]
  GemDownload Pluck (0.7ms)  SELECT "gem_downloads"."rubygem_id", "gem_downloads"."count" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" IN ($1, $2, $3, $4) AND "gem_downloads"."version_id" = $5 ORDER BY "gem_downloads"."rubygem_id" ASC  [["rubygem_id", 9], ["rubygem_id", 10], ["rubygem_id", 11], ["rubygem_id", 8], ["version_id", 0]]
  GemDownload Update All (1.0ms)  UPDATE "gem_downloads" SET count = "gem_downloads".count + updates_by_gem.downloads FROM (SELECT UNNEST(ARRAY[9,10,11,8]) AS r_id, UNNEST(ARRAY[0,0,0,0]) AS downloads) AS updates_by_gem WHERE "gem_downloads".rubygem_id = updates_by_gem.r_id AND "gem_downloads".version_id = 0
  GemDownload Load (0.5ms)  UPDATE "gem_downloads" SET count = count + 0 WHERE id = (SELECT "gem_downloads"."id" FROM "gem_downloads" WHERE "gem_downloads"."rubygem_id" = 0 AND "gem_downloads"."version_id" = 0) RETURNING *
```